Move groupShippingRatesByMethodPriceCurrency into local component folder.
We don't need it to be in utils folder, and we want to reuse the typedef in local folder.
