+++ js/src/components/shipping-rate-section/estimated-shipping-rates-card/estimated-shipping-rates-card.js
@@ -12,7 +12,7 @@ import Section from '.~/wcdl/section';
 import AppButtonModalTrigger from '.~/components/app-button-modal-trigger';
 import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import useStoreCurrency from '.~/hooks/useStoreCurrency';
-import groupShippingRatesByMethodPriceCurrency from '.~/utils/groupShippingRatesByMethodPriceCurrency';
+import groupShippingRatesByMethodPriceCurrency from './groupShippingRatesByMethodPriceCurrency';
 import ShippingRateInputControl from './shipping-rate-input-control';
 import AddRateFormModal from './add-rate-form-modal';
 import { SHIPPING_RATE_METHOD } from '.~/constants';
