+++ js/src/components/shipping-rate-section/estimated-shipping-rates-card/groupShippingRatesByMethodPriceCurrency.test.js
